HOI3 Custom Music Modding Tool
This tool writes a songs.txt file for HOI3 in a specified folder. It is written in Python and is converted to an executable using auto-py-to-exe. 

It's incredibly bare but it should be simple enough to use. It reads every file in the directory and adds the file name of files ending in .mp3 into songs.txt.